refactor model laoder
